fix(activate): prepare and pass env vars to serverless-api activateBuild
Fixes: #109
